Atualiza hwingui.h adiciona casts nas macros hwg_ret_LRESULT e hwg_vmPushWPARAM

+++ include/hwingui.h
@@ -158,7 +158,7 @@ HB_EXTERN_END
 
 #define hwg_ret_BOOL(x) hb_retl(x)
 #define hwg_ret_HWND(n) HB_RETHANDLE(n)
-#define hwg_ret_LRESULT(x) hb_retnl(x)
+#define hwg_ret_LRESULT(x) hb_retnl((long)x)
 //#define hwg_ret_PAINTSTRUCT(n) HB_RETHANDLE(n)
 #define hwg_ret_PAINTSTRUCT(n) hb_retptr(n)
 #define hwg_ret_HBRUSH(x) HB_RETHANDLE(x)
@@ -178,5 +178,5 @@ HB_EXTERN_END
 #define hwg_ret_COLORREF(x) hb_retnl(x)
 
 #define hwg_vmPushUINT(x) hb_vmPushInteger(x)
-#define hwg_vmPushWPARAM(x) hb_vmPushLong(x)
+#define hwg_vmPushWPARAM(x) hb_vmPushLong((long)x)
 #define hwg_vmPushLPARAM(x) HB_PUSHITEM(x)
